Add files via upload
This is an edited matlab script that connects to both devices of the parallel configuration for data acquisition from the gatt characteristic.  The devices will start to scan after running the script and matlab successfully connects to each device.  Please change the hardcoded bt addresses to your devices' addresses, the characteristic UUID will remain the same if using the gatt.xml provided in the parallel folder